Implemented code, so user can't schedule at an illogical time.
I added the following changes:

Date and time checking so that a user is not able to schedule at an illogical time, only for the current date. So for example if user is viewing availability for technician on Oct 26 and that is today's date and it is 2 in the afternoon, the user would not be able to schedule a time at 9:00am.

Implemented a small change to the calendar.html to display the date that the user has selected.